Renamed api > functions and models > classes for simplicity
