+++ shui/classes/file_with_hash.py
