+++ shui/classes/fileinfo.py
