+++ shui/functions/__init__.py
@@ -1,4 +1,4 @@
-"""Module for interacting with remote Apache server"""
+"""Module for functions that interact with the remote Apache server"""
 from .download import download_version, get_file_details
 from .install import extract_tarball
 from .versions import get_versions
